Added jUnit Support
git-svn-id: https://openwms2005.svn.sourceforge.net/svnroot/openwms2005@24 fda5f131-174c-0410-8613-e0905860106b

+++ OpenWMS-common/trunk/src/org/openwms/domain/common/TypePlacingRule.java
@@ -59,7 +59,7 @@ public class TypePlacingRule implements Serializable {
 	 * <code>LocationType</code> the privilegeLevel must be set to -1.
 	 */
 	@Column(name = "PRIVILEGE_LEVEL", nullable = false)
-	private short privilegeLevel = 0;
+	private int privilegeLevel = 0;
 
 	/**
 	 * The allowed <code>LocationType</code> on which the owning
@@ -69,7 +69,7 @@ public class TypePlacingRule implements Serializable {
 	private LocationType allowedLocationType;
 
 	/* ----------------------------- methods ------------------- */
-	public TypePlacingRule(short privilegeLevel, LocationType allowedLocationType) {
+	public TypePlacingRule(int privilegeLevel, LocationType allowedLocationType) {
 		super();
 		this.privilegeLevel = privilegeLevel;
 		this.allowedLocationType = allowedLocationType;
@@ -103,7 +103,7 @@ public class TypePlacingRule implements Serializable {
 	 * 
 	 * @return the privilegeLevel.
 	 */
-	public short getPrivilegeLevel() {
+	public int getPrivilegeLevel() {
 		return privilegeLevel;
 	}
